(PC-17816) refactor(profile): remove reusable folder
