--- src/features/profile/components/reusables.ts
@@ -1,39 +1,0 @@
-import { StyleSheet } from 'react-native'
-import styled from 'styled-components/native'
-
-import { getSpacing, Typo } from 'ui/theme'
-
-export const ProfileContainer = styled.View(({ theme }) => ({
-  flex: 1,
-  backgroundColor: theme.colors.white,
-  flexDirection: 'column',
-  paddingHorizontal: getSpacing(6),
-}))
-
-export const ScrollableProfileContainer = styled.ScrollView(({ theme }) => ({
-  flex: 1,
-  flexDirection: 'column',
-  backgroundColor: theme.colors.white,
-  paddingHorizontal: getSpacing(6),
-}))
-
-export const GreyContainer = styled.View(({ theme }) => ({
-  backgroundColor: theme.colors.greyLight,
-  borderRadius: 6,
-  width: '100%',
-}))
-
-export const accordionStyle = StyleSheet.create({
-  title: {
-    paddingHorizontal: getSpacing(4),
-    paddingBottom: getSpacing(4),
-  },
-  body: {
-    paddingHorizontal: getSpacing(4),
-    paddingBottom: getSpacing(4),
-  },
-})
-
-export const Description = styled(Typo.Body)({
-  textAlign: 'left',
-})
