+++ src/features/profile/pages/Profile.tsx
@@ -9,7 +9,7 @@ import { useAuthContext } from 'features/auth/AuthContext'
 import { useLogoutRoutine } from 'features/auth/logout/useLogoutRoutine'
 import { useFavoritesState } from 'features/favorites/context/FavoritesWrapper'
 import { ProfileHeader } from 'features/profile/components/Header/ProfileHeader/ProfileHeader'
-import { ProfileContainer } from 'features/profile/components/reusables'
+import { ProfileContainer } from 'features/profile/components/PageProfileSection/ProfileContainer'
 import { SectionWithSwitch } from 'features/profile/components/SectionWithSwitch/SectionWithSwitch'
 import { SocialNetwork } from 'features/profile/components/SocialNetwork/SocialNetwork'
 import { env } from 'libs/environment'
